requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
openai-whisper==20240930
spacy==3.7.5
